write protect disable is working
Software write protect can be disabled by writing a
special sequence of data values to addresses $5555
and $2AAA --- however, the data bytes must be written
within 150 us of each other, and it is not possible
to meet this timing requirement when sending data to
the 74HC595 shift registers using digitalWrite.
Rewriting setAddr and setData to directly bit-bang the
underlying AVR I/O ports allows the firmware to meeting
the timing requirements.  Successfully wrote a byte of
data to an AT28C256 which was write protected.

As an aside, the oscilloscope was incredibly helpful
for seeing the exact timing of the CE and WE signals.
It was immediately obvious that the pulses for each
data byte written were too far apart (more than 400 us).